Reduce the lag when changing screens
