+++ tools-ui/src/main/java/com/hedera/hashgraph/client/ui/utilities/Utilities.java
@@ -23,6 +23,9 @@ import com.codahale.passpol.PasswordPolicy;
 import com.codahale.passpol.Status;
 import com.hedera.hashgraph.client.core.exceptions.HederaClientException;
 import com.hedera.hashgraph.client.core.json.Timestamp;
+import com.hedera.hashgraph.client.core.utils.EncryptionUtils;
+import com.hedera.hashgraph.client.ui.Controller;
+import com.hedera.hashgraph.sdk.AccountInfo;
 import com.hedera.hashgraph.sdk.Hbar;
 import javafx.animation.PauseTransition;
 import javafx.scene.control.Alert;
@@ -44,12 +47,16 @@ import java.text.DateFormat;
 import java.text.DecimalFormat;
 import java.text.DecimalFormatSymbols;
 import java.text.SimpleDateFormat;
+import java.util.ArrayList;
+import java.util.Collections;
 import java.util.Date;
+import java.util.List;
 import java.util.TimeZone;
 
 import static com.hedera.hashgraph.client.core.constants.Constants.GREEN_STYLE;
 import static com.hedera.hashgraph.client.core.constants.Constants.MAX_PASSWORD_LENGTH;
 import static com.hedera.hashgraph.client.core.constants.Constants.MIN_PASSWORD_LENGTH;
+import static com.hedera.hashgraph.client.core.constants.Constants.PUB_EXTENSION;
 import static com.hedera.hashgraph.client.core.constants.Constants.RED_STYLE;
 
 public class Utilities {
@@ -323,6 +330,7 @@ public class Utilities {
 				break;
 		}
 	}
+
 	public static void setupCharacterCount(PasswordField recoverAppPasswordField, Label recoverCharacterCount,
 			ImageView recoverCheckPassword, Label recoverPasswordErrorLabel,
 			PasswordField recoverReEnterPasswordField) {
@@ -342,5 +350,28 @@ public class Utilities {
 			recoverPasswordErrorLabel.setVisible(true);
 			recoverReEnterPasswordField.setDisable(true);
 		}
+	}
+
+	/**
+	 * Given an account info, returns a list of string keys
+	 *
+	 * @param info
+	 * 		the account info
+	 * @param controller
+	 * 		the controller
+	 * @return a List of strings: If the controller has information about the public key, it uses the nickname,
+	 * otherwise
+	 * 		it shows the complete hex.
+	 */
+	public static List<String> getKeysFromInfo(AccountInfo info, Controller controller) {
+		var flatKey = EncryptionUtils.flatPubKeys(Collections.singletonList(info.key));
+		List<String> knownKeys = new ArrayList<>();
+		for (var key : flatKey) {
+			var keyName = controller.showKeyString(key);
+			if (keyName.endsWith(PUB_EXTENSION)) {
+				knownKeys.add(keyName);
+			}
+		}
+		return knownKeys;
 	}
 }
